fix: recompute DeepSeek V4 hash routing weights
Use the hash routing custom op only for selected expert ids in the sqrtsoftplus path, then recompute top-k weights from router logits in FP32 before renormalization.

This avoids propagating NaN weights returned by the custom op while keeping the routing ids unchanged. Keep W4A8 MoE top-k weights in FP32 before expert scaling and add regression coverage for comm-aligned input ids and NaN weight recomputation.

+++ vllm_ascend/quantization/methods/w4a8.py
@@ -386,7 +386,9 @@ class AscendW4A8DynamicFusedMoEMethod(AscendMoEScheme):
             random_matrix = torch.rand(topk_ids.size(0), num_logical_experts, device=topk_ids.device)
             topk_ids = torch.argsort(random_matrix, dim=1)[:, : topk_ids.size(1)].to(topk_ids.dtype)
 
-        topk_weights = topk_weights.to(x.dtype)
+        # Keep router weights in FP32 for expert scaling. Casting to BF16 here
+        # loses precision before MC2/ALLTOALL combine can promote them again.
+        topk_weights = topk_weights.to(torch.float32)
 
         if self.dynamic_eplb:
             w1 = layer.w13_weight_list
